refactor: Singleton sortable
